refactor(ArticleList): move to _features/articles directory
