+++ apps/web/app/_features/articles/index.ts
@@ -10,3 +10,5 @@ export type { ArticleMeta, TagEnum } from "./type";
 export { tagList, tagsSchema } from "./type";
 /** @public */
 export { tagLabelMap } from "./constants";
+/** @public */
+export { ArticleList } from "./ArticleList";
