--- apps/web/app/articles/_features/index.ts
@@ -1,1 +1,0 @@
-export { ArticleList } from "./ArticleList";
